code/requirements.txt
torch>=2.0.0,<3.0.0
torchvision>=0.14.0
boto3>=1.20.18,<2.0.0
scikit-learn>=1.0.0,<2.0.0
Pillow>=8.3.2